+++ packages/util/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [9.12.0](https://github.com/dereekb/dbx-components/compare/v9.11.13-dev...v9.12.0) (2022-11-07)
+
+
+### Features
+
+* added @dereekb/util/fetch ([6afa4a4](https://github.com/dereekb/dbx-components/commit/6afa4a48df62791a4b9ee8da67cb8e8bd00d3bcc))
+
+
+
 ## [9.11.13](https://github.com/dereekb/dbx-components/compare/v9.11.12-dev...v9.11.13) (2022-11-05)